Collections must use ArrayList instay of simple arrays.

+++ src/perceptron_and/Neuron.java
@@ -5,20 +5,20 @@ package perceptron_and;
  * @author Iulisloi Zacarias
  */
 public class Neuron {
-    
-    private double value;
+
+    private int value;
     private double weight;
-    
+
     public Neuron() {
         this.value = 0;
         this.weight = 0;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(int value) {
         this.value = value;
     }
 
@@ -29,5 +29,5 @@ public class Neuron {
     public void setWeight(double weight) {
         this.weight = weight;
     }
-    
+
 }
